Fix build warnings and errors on clang shared build.

Clang seems to need MAPTK_CORE_EXPORT on algorithm class definitions
that are essentially header only.  I think the issue is that the
CRTP base class is not header only and needs to be instantiated.
So there are .cxx files for these headers, but they contain only
base class instantiations.  Adding MAPTK_CORE_EXPORT to the CRTP base class
is not sufficient.

Fixed build warnings on VS10 shared build.

* remotes/origin/dev/fix-warnings-errors:
  fixed typos
  updated release notes
  address Windows VS10 warnings
  fixed indentation
  export algorithm class definitions
  avoid unused variable warning

+++ maptk/ocv/estimate_homography.cxx
@@ -1,5 +1,5 @@
 /*ckwg +5
- * Copyright 2013 by Kitware, Inc. All Rights Reserved. Please refer to
+ * Copyright 2013-2014 by Kitware, Inc. All Rights Reserved. Please refer to
  * KITWARE_LICENSE.TXT for licensing information, or contact General Counsel,
  * Kitware, Inc., 28 Corporate Drive, Clifton Park, NY 12065.
  */
@@ -34,11 +34,13 @@ estimate_homography
   std::vector<cv::Point2f> points1, points2;
   BOOST_FOREACH(const vector_2d& v, pts1)
   {
-    points1.push_back(cv::Point2f(v.x(), v.y()));
+    points1.push_back(cv::Point2f(static_cast<float>(v.x()),
+                                  static_cast<float>(v.y())));
   }
   BOOST_FOREACH(const vector_2d& v, pts2)
   {
-    points2.push_back(cv::Point2f(v.x(), v.y()));
+    points2.push_back(cv::Point2f(static_cast<float>(v.x()),
+                                  static_cast<float>(v.y())));
   }
 
   cv::Mat inliers_mat;
